requirements.txt
aiofiles==23.2.1
aiogram==2.23.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
exceptiongroup==1.2.1
executor==23.2
fasteners==0.19
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
pillow==10.3.0
property-manager==3.0
psycopg2-binary==2.9.9
pyreadline3==3.4.1
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.9.0
verboselogs==1.7
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4